requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.2
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cryptography==3.4.8
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
ebaysdk==2.1.5
executing==1.2.0
freezegun==0.3.15
frozenlist==1.3.3
gevent==20.9.0
greenlet==0.4.17
idna==2.8
ipdb==0.13.11
ipython==8.11.0
isodate==0.6.1
jedi==0.18.2
Jinja2==2.11.3
libsass==0.18.0
lxml==4.6.5
magic-filter==1.0.9
MarkupSafe==1.1.0
matplotlib-inline==0.1.6
multidict==6.0.4
num2words==0.5.6
ofxparse==0.19
parso==0.8.3
passlib==1.7.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
polib==1.1.0
prompt-toolkit==3.0.38
psutil==5.6.7
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.1
Pygments==2.14.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyserial==3.4
python-dateutil==2.7.3
python-ldap==3.4.0
python-stdnum==1.13
pytz==2019.3
pyusb==1.0.2
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
urllib3==1.26.5
vobject==0.9.6.1
wcwidth==0.2.6
Werkzeug==0.16.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarl==1.8.2
zeep==3.4.0
zope.event==4.6
zope.interface==5.5.2
